Refactoring some style definitions to avoid duplicated code
