+++ webapp/src/components/lifelines/GraphComponent.js
@@ -1,5 +1,4 @@
 import React, { useEffect, useState } from 'react';
-import { Box, Typography } from '@mui/material';
 import { BarChart, Bar, XAxis, YAxis, CartesianGrid, Tooltip, ResponsiveContainer, LabelList } from 'recharts';
 
 const GraphComponent = ({ correctAnswer, distractors }) => {
