+++ webapp/src/components/selections/GameTopicSelection.js
@@ -8,98 +8,7 @@ import {
   LocationCity, Flag, SportsBasketball, MusicNote, Public, Sports, Map, Event, PushPin,
   Piano, Female, SportsSoccer, Language, SportsMotorsports, People, Landscape
 } from "@mui/icons-material"
-import useAxios from "../../hooks/useAxios"
-
-const StyledContainer = styled(Container)(({ theme }) => ({
-  padding: theme.spacing(4),
-  display: "flex",
-  flexDirection: "column",
-  alignItems: "center",
-  gap: theme.spacing(3),
-}))
-
-const SectionPaper = styled(Paper)(({ theme }) => ({
-  padding: theme.spacing(3),
-  width: "100%",
-  background: theme.palette.background.paper,
-  borderRadius: theme.shape.borderRadius,
-  boxShadow: theme.shadows[2],
-}))
-
-const SectionTitle = styled(Typography)(({ theme }) => ({
-  marginBottom: theme.spacing(2),
-  fontWeight: "bold",
-  color: theme.palette.primary.main,
-  textAlign: "center",
-  textTransform: "uppercase",
-  borderBottom: `2px solid ${theme.palette.primary.main}`,
-  paddingBottom: theme.spacing(1),
-}))
-
-const StyledButton = styled(Button)(({ theme }) => ({
-  marginTop: theme.spacing(2),
-  padding: theme.spacing(1.5, 4),
-  fontSize: "1.2rem",
-  fontWeight: "bold",
-  borderRadius: 30,
-  background: theme.palette.gradient.main.right,
-  "&:hover": {
-    background: theme.palette.gradient.hover.right,
-    transform: "scale(1.03)",
-    color: "white"
-  },
-  "&.Mui-disabled": {
-    background: theme.palette.background.paper,
-    color: theme.palette.text.disabled,
-  },
-  transition: theme.transitions.create(["transform"], {
-    duration: theme.transitions.duration.short,
-  }),
-}))
-
-
-const TopicOption = styled(FormControlLabel, {
-  shouldForwardProp: (prop) => prop !== "isSelected",
-})(({ theme, isSelected }) => ({
-  marginTop: theme.spacing(1),
-  marginBottom: theme.spacing(1),
-  padding: theme.spacing(1),
-  borderRadius: theme.shape.borderRadius,
-  background: isSelected ? theme.palette.gradient.main.left : theme.palette.background.default,
-  color: isSelected ? theme.palette.common.white : theme.palette.text.primary,
-  width: "100%",
-  margin: 0,
-  "& .MuiFormControlLabel-label": {
-    fontWeight: "bold",
-  },
-}))
-
-
-const TopicButton = styled(Button, {
-  shouldForwardProp: (prop) => prop !== "isSelected",
-})(({ theme, isSelected }) => ({
-  padding: theme.spacing(1.5, 2),
-  borderRadius: theme.shape.borderRadius,
-  fontWeight: "bold",
-  background: isSelected ? theme.palette.gradient.main.left : theme.palette.background.default,
-  color: isSelected ? theme.palette.common.white : theme.palette.text.primary,
-  border: isSelected ? "none" : `1px solid ${theme.palette.divider}`,
-  boxShadow: isSelected ? theme.shadows[3] : "none",
-  textTransform: "none",
-  transition: theme.transitions.create(["background", "transform", "box-shadow"], {
-    duration: theme.transitions.duration.short,
-  }),
-  width: "100%",
-  height: "100%",
-  "&:hover": {
-    background: isSelected ? theme.palette.gradient.hover.left : theme.palette.action.hover,
-    transform: "translateY(-2px)",
-    boxShadow: isSelected ? theme.shadows[4] : theme.shadows[1],
-  },
-  "& .MuiButton-startIcon": {
-    marginRight: theme.spacing(1),
-  },
-}))
+import { StyledContainer, SectionPaper, SectionTitle, StyledButton, ModeButton, TopicButton } from './SelectionStyles'
 
 const CATEGORY_MAP = {
   "Geography": [
